test: refactor mock utils, added new one
