+++ src/accounts/accounts.utils.spec.ts
@@ -15,8 +15,7 @@ import { TOKENS } from '../config/tokens.config';
 import { LeaderBoardType } from '../leaderboards/enums/leaderboard-type.enum';
 import { UserClaimMetadata } from '../rewards/entities/user-claim-metadata';
 import { MOCK_VAULT_DEFINITION, TEST_ADDR, TEST_CURRENT_BLOCK } from '../test/constants';
-import { mockBalance, mockQuery, setupMockChain } from '../test/mocks.utils';
-import { defaultAccount, randomSnapshot } from '../test/tests.utils';
+import { defaultAccount, mockBalance, mockQuery, randomSnapshot,setupMockChain } from '../test/mocks.utils';
 import { fullTokenMockMap } from '../tokens/mocks/full-token.mock';
 import * as vaultsUtils from '../vaults/vaults.utils';
 import { getAccounts, getCachedBoost, getLatestMetadata, queryCachedAccount, toVaultBalance } from './accounts.utils';
