+++ src/indexers/leaderboard-indexer.spec.ts
@@ -4,8 +4,7 @@ import * as accountsUtils from '../accounts/accounts.utils';
 import { CachedBoost } from '../aws/models/cached-boost.model';
 import { TOKENS } from '../config/tokens.config';
 import { BoostData } from '../rewards/interfaces/boost-data.interface';
-import { mockBatchDelete, mockBatchPut, mockQuery, setupMockChain } from '../test/mocks.utils';
-import { randomCachedBoosts } from '../test/tests.utils';
+import { mockBatchDelete, mockBatchPut, mockQuery, randomCachedBoosts, setupMockChain } from '../test/mocks.utils';
 import { indexBoostLeaderBoard } from './leaderboard-indexer';
 
 describe('leaderboard-indexer', () => {
