+++ src/test/mocks.utils/contracts/mock.contract.base.ts
@@ -1,0 +1,10 @@
+import { mock } from "jest-mock-extended";
+
+export function mockContract<T>(factory: any) {
+  const mockContract = mock<T>();
+
+  // @ts-ignore
+  jest.spyOn(factory, 'connect').mockImplementation(async () => mockContract);
+
+  return mockContract;
+}
